refactor: rename upload.cpp to network.cpp
